ceph_iscsi_config: disable emulate_legacy_capacity
if the corresponding control is set.

The controls are applied after enabling the backstore, and in the
case of emulate_legacy_capacity control it is too late, as it is
checked by the backend on enabling. Thus we need to provide this
setting earlier.

Note, this requires a version of rtslib that supports
disable_emulate_legacy_capacity param.

Signed-off-by: Mykola Golub <mgolub@suse.com>